requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
streamlit
python-dotenv
pydantic
email-validator
